Enhance run_all_evals.sh to include optimization and evaluation steps for Executorch; update scripts for dynamic output paths and parameter handling

+++ utils/quant_utils.py
@@ -356,7 +356,7 @@ class WeightQuantizer(torch.nn.Module):
         if self.mse:
             best = torch.full(
                 [x.shape[0], x.shape[1]], float("inf"), device=x.device
-            ).type_as(x)
+            )
             for i in range(int(self.maxshrink * self.grid)):
                 p = 1 - i / self.grid
                 xmin1 = p * xmin
